Renamed simple_service and run_simple_service according to the new scheme
