+++ examples/run_simple_service.cc
@@ -1,57 +1,45 @@
 /*
- * run_simple_service.cc
+ * simple_service.cc
  *
  *  Created on: Aug 23, 2018
  *      Author: N.S. Oblath
  */
 
-#define DRIPLINE_EXAMPLES_API_EXPORTS
-
+#include "agent_config.hh"
+#include "dripline_constants.hh"
+#include "dripline_version.hh"
 #include "run_simple_service.hh"
 
-#include "dripline_error.hh"
-
+#include "application.hh"
 #include "logger.hh"
-#include "signal_handler.hh"
 
-#include <chrono>
-#include <thread>
+using namespace dripline;
 
-LOGGER( dlog, "run_simple_service" )
+LOGGER( dlog, "simple_service" );
 
-namespace dripline
+int main( int argc, char** argv )
 {
+    scarab::main_app the_main;
 
-    run_simple_service::run_simple_service( const scarab::param_node& a_config ) :
-            service( a_config, "simple" ),
-            f_return( RETURN_SUCCESS )
-    {
-    }
+    the_main.set_version( new dripline::version() );
 
-    run_simple_service::~run_simple_service()
-    {
-    }
+    the_main.default_config().add( "amqp", amqp_config() );
 
-    void run_simple_service::execute()
-    {
-        scarab::signal_handler t_sig_hand;
-        t_sig_hand.add_cancelable( this );
+    add_amqp_options( the_main );
 
-        try
-        {
-            if( ! start() ) throw dripline_error() << "Unable to start service";
+    int the_return = -1;
 
-            if( ! listen() ) throw dripline_error() << "Unable to start listening";
+    auto t_service_callback = [&](){
+        auto the_service = std::make_shared< run_simple_service >( the_main.master_config()["amqp"].as_node() );
 
-            if( ! stop() ) throw dripline_error() << "Unable to stop service";
-        }
-        catch( std::exception& e )
-        {
-            LERROR( dlog, "Exception caught: " << e.what() );
-            f_return = RETURN_ERROR;
-        }
+        the_service->execute();
 
-        return;
-    }
+        the_return = the_service->get_return();
+    };
 
-} /* namespace dripline */
+    the_main.callback( t_service_callback );
+
+    CLI11_PARSE( the_main, argc, argv );
+
+    return the_return;
+}
